requirements.txt
Django==2.0.5
graphene-framework==0.0.4
graphene-django==2.1.0